app/requirements.txt
django==3.0
djangorestframework
django-filter  # Filtering support
markdown       # Markdown support for the browsable API.
django-mssql-backend
# mysqlclient
django-webpack-loader
gunicorn 
python-dotenv
django-rest-framework-social-oauth2